Create API_Consumer class
This commit create the API_Consumer class to work with the Yahoo API.

The class has two class methods as of now:
- `#get_forecast` with makes a call to the Yahoo API_Consumer
- `#parse_forecast` which sifts through the Yahoo response for the
forecast data

Currently the logic that belongs in `#parse_forecast` lives in the
controller instead of the model. That will be addressed in the next
commit.

:notebook: Remember to look closely at API responses. I spent 25
minutes debugging my controller before realizing that the response used
strings for keys instead of the symbols I'd been typing.