Change #parse_forecast output distinguish today
This commit changes the output for the parse_forecast. It splits it into
a hash with 2 keys: `:today` and `:next_nine_days`. The goal here it to
minimize the client's work in understanding the method's response.

This commit also renames `#get_forecast` to `#query_weather_api` and
makes it a private method.