Build the Day and Forecast models; Remove AR
This commit adds attributes to the forecast and day models. It removes
ActiveRecord inheritance since nothing lives in the database and creates
days as an array collection within a forecast.

Additionally, because a forecast now has associated days, its methods
have been converted to instance methods that act on its `@query`
instance variable.