Implement sticky forecast
This commit adds conditional logic to the get request on `/` that
loads and appends the forecast for the last city someone selected,
provided they have a value stored in `session[:current_city]`.

It also adds some conditional logic to the index view to load the
forecast if necessary and Javascript updates to handle repeat requests.